+++ src/lib/errors/max-array-length.ts
@@ -4,7 +4,7 @@ export namespace MaxArrayLengthError {
   export type Name = "MaxArrayLength";
   export const name: Name = "MaxArrayLength";
   export interface Data {
-    array: any[];
+    array: ArrayLike<any>;
     max: number;
   }
   export type Cause = undefined;
@@ -12,7 +12,7 @@ export namespace MaxArrayLengthError {
   export function format({array, max}: Data): string {
     return `Expected array length (${array.length}) to be less than or equal to ${max}`;
   }
-  export function create(array: any[], max: number): Type {
+  export function create(array: ArrayLike<any>, max: number): Type {
     return Incident(name, {array, max}, format);
   }
 }