+++ src/lib/errors/min-array-length.ts
@@ -4,7 +4,7 @@ export namespace MinArrayLengthError {
   export type Name = "MinArrayLength";
   export const name: Name = "MinArrayLength";
   export interface Data {
-    array: any[];
+    array: ArrayLike<any>;
     min: number;
   }
   export type Cause = undefined;
@@ -12,7 +12,7 @@ export namespace MinArrayLengthError {
   export function format({array, min}: Data): string {
     return `Expected array length (${array.length}) to be greater than or equal to ${min}`;
   }
-  export function create(array: any[], min: number): Type {
+  export function create(array: ArrayLike<any>, min: number): Type {
     return Incident(name, {array, min}, format);
   }
 }